requirements.txt
django==5.2.2
djangorestframework==3.16.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.11.1
